feat(osn/api): Upstash REST Redis backend for Workers (migration P2)
Add a Workers-compatible Redis backend selected at composition time, without
changing any consumer call sites and without dragging ioredis into the Workers
bundle.

- Split ioredis to the `@shared/redis/ioredis` subpath: wrapIoRedis,
  createClientFromUrl, ConnectableRedisClient, and the Effect RedisLive layer
  no longer live in the top-level entry's import graph. The `@shared/redis`
  entry now exports only the RedisClient interface, the in-memory client, and
  the new Upstash client — no static ioredis import (verified).
- Add `@shared/redis/upstash`: wrapUpstash + createUpstashClient. get() uses
  automaticDeserialization:false for raw-string round-trips; set() maps pxMs to
  { px }; eval() passes through (numeric/`"1"` returns preserved); quit() is a
  no-op for the stateless REST transport.
- Add `initRedisClientFromEnv(env)` to osn/api: synchronous, ioredis-free,
  side-effect-free selector — Upstash when both UPSTASH_REDIS_REST_URL and
  UPSTASH_REDIS_REST_TOKEN are set, else in-memory. No health check /
  REDIS_REQUIRED / process.exit (those stay on the Bun initRedisClient path).
- Repoint the Bun-only importers (osn/api + pulse/api redis.ts and their tests)
  at `@shared/redis/ioredis`. Bun path unchanged.

Tests: shared/redis 52 pass (incl. new ioredis + upstash contract tests);
osn/api 614 pass; pulse/api 502 pass. shared/redis + osn/api + pulse/api
type-checks clean.
