+++ shared/redis/src/service.ts
@@ -1,24 +1,24 @@
 /**
  * Effect-based Redis service for lifecycle management.
  *
- * - `RedisLive` — connects via REDIS_URL; Layer.scoped finalizer calls quit()
  * - `RedisMemoryLive` — in-memory fallback for dev/test (no Redis server)
+ *
+ * The ioredis-backed `RedisLive` layer lives in `./ioredis` (re-exported from
+ * the `@shared/redis/ioredis` subpath) so this module — and the top-level
+ * `@shared/redis` entry — stay free of any static `ioredis` import and remain
+ * loadable on Cloudflare Workers.
  */
 
 import { Context, Effect, Layer } from "effect";
-import IORedis from "ioredis";
 
 import type { RedisClient } from "./client";
-import { wrapIoRedis, createMemoryClient } from "./client";
-import { RedisError } from "./errors";
+import { createMemoryClient } from "./client";
 
 export interface RedisService {
   readonly client: RedisClient;
 }
 
 export class Redis extends Context.Tag("@shared/redis/Redis")<Redis, RedisService>() {}
-
-const STARTUP_PING_TIMEOUT_MS = 5_000;
 
 /** Redact credentials from Redis URLs in error messages (S-M3). */
 function sanitizeCause(cause: unknown): string {
@@ -28,53 +28,6 @@ function sanitizeCause(cause: unknown): string {
     return `${scheme}[REDACTED]@`;
   });
 }
-
-/**
- * Live layer — connects to `REDIS_URL`. `Layer.scoped` ensures `quit()` on
- * shutdown. Fails with `RedisError` if `REDIS_URL` is unset or the connection
- * cannot be verified via PING.
- */
-export const RedisLive: Layer.Layer<Redis, RedisError> = Layer.scoped(
-  Redis,
-  Effect.gen(function* () {
-    const url = process.env.REDIS_URL;
-    if (!url) {
-      return yield* Effect.fail(
-        new RedisError({ cause: "REDIS_URL environment variable is not set" }),
-      );
-    }
-
-    // S-M1: warn when production connection is unencrypted
-    if (!url.startsWith("rediss://") && process.env.NODE_ENV === "production") {
-      yield* Effect.logWarning(
-        "REDIS_URL does not use TLS (rediss://) — connection is unencrypted",
-      );
-    }
-
-    const raw = new IORedis(url);
-    const client = wrapIoRedis(raw);
-
-    // P-I2: startup ping with timeout to prevent indefinite hangs
-    let timer: ReturnType<typeof setTimeout>;
-    yield* Effect.tryPromise({
-      try: () =>
-        Promise.race([
-          client.ping().finally(() => clearTimeout(timer)),
-          new Promise<never>((_, reject) => {
-            timer = setTimeout(
-              () => reject(new Error("Redis startup ping timed out")),
-              STARTUP_PING_TIMEOUT_MS,
-            );
-          }),
-        ]),
-      catch: (cause) => new RedisError({ cause: sanitizeCause(cause) }),
-    }).pipe(Effect.tapError(() => Effect.logError("Redis connection failed")));
-
-    yield* Effect.addFinalizer(() => Effect.promise(() => client.quit().catch(() => {})));
-
-    return { client };
-  }),
-);
 
 /**
  * In-memory layer for dev/test — no Redis server required.
